[Debug] added the component (closes #6828, closes #6834, closes #7330)
