+++ src/Symfony/Component/Debug/.gitignore
@@ -1,0 +1,5 @@
+vendor/
+composer.lock
+phpunit.xml
+Tests/ProjectContainer.php
+Tests/classes.map
