+++ src/Symfony/Component/Debug/CHANGELOG.md
@@ -1,0 +1,7 @@
+CHANGELOG
+=========
+
+2.3.0
+-----
+
+ * added the component
