+++ src/Symfony/Component/Debug/Exception/FatalErrorException.php
@@ -9,17 +9,13 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\HttpKernel\Exception;
-
-use Symfony\Component\Debug\Exception\FatalErrorException as DebugFatalErrorException;
+namespace Symfony\Component\Debug\Exception;
 
 /**
  * Fatal Error Exception.
  *
  * @author Konstanton Myakshin <koc-dp@yandex.ru>
- *
- * @deprecated Deprecated in 2.3, to be removed in 3.0. Use the same class from the Debug component instead.
  */
-class FatalErrorException extends DebugFatalErrorException
+class FatalErrorException extends \ErrorException
 {
 }
